src/scraper.py
import requests
from bs4 import BeautifulSoup
from urllib.parse import urlparse
import click

class GolfCourseScraper:
    def __init__(self):
        self.headers = {
            'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
        }

    def validate_url(self, url):
        """Validate if the URL is properly formatted."""
        try:
            result = urlparse(url)
            return all([result.scheme, result.netloc])
        except:
            return False

    def fetch_page(self, url):
        """Fetch the webpage content."""
        if not self.validate_url(url):
            raise ValueError("Invalid URL format")

        try:
            response = requests.get(url, headers=self.headers)
            response.raise_for_status()
            return response.text
        except requests.RequestException as e:
            raise click.ClickException(f"Failed to fetch the webpage: {str(e)}")

    def parse_content(self, html_content):
        """Parse the HTML content and extract relevant information."""
        soup = BeautifulSoup(html_content, 'lxml')
        
        # This is a basic implementation that just returns the text content
        # You'll need to customize this based on the specific golf course website structure
        content = soup.get_text(separator='\n', strip=True)
        
        return content

def scrape_golf_course(url):
    """Main function to scrape a golf course website."""
    scraper = GolfCourseScraper()
    try:
        html_content = scraper.fetch_page(url)
        content = scraper.parse_content(html_content)
        return content
    except Exception as e:
        raise click.ClickException(str(e)) 